+++ CHANGES.rst
@@ -1,10 +1,19 @@
 Changelog
 *********
 
-0.1.1 (unreleased)
+0.2.1 (unreleased)
 ==================
 
 - Nothing changed yet.
+
+
+0.2.0 (2018-01-02)
+==================
+
+- no more open browser by default
+  pytest-play is a generic test engine and it could be used for non UI tests too.
+
+  So there is no need to open the browser for non UI tests (eg: API tests)
 
 
 0.1.0 (2017-12-22)